Add filepicker_field_tag helper method
Add shared example spec to test filepicker_field and filepicker_field_tag
