+++ app/helpers/filepicker_rails/form_helper.rb
@@ -1,13 +1,7 @@
 module FilepickerRails
   module FormHelper
 
-    FILEPICKER_OPTIONS_TO_DASHERIZE = [:button_text, :button_class, :mimetypes,
-                                       :extensions, :container, :services,
-                                       :drag_text, :drag_class, :store_path,
-                                       :store_location, :store_access,
-                                       :store_container, :multiple]
-
-    FILEPICKER_OPTIONS_TO_CAMELIZE = [:max_size, :max_files, :open_to]
+    include FilepickerRails::Tag
 
     # Creates a filepicker field, accepts optional `options` hash for configuration.
     #
@@ -58,16 +52,7 @@ module FilepickerRails
 
     private
 
-      attr_reader :input_options, :method, :type, :object_name, :template
-
-      def define_input_options(options)
-        @type = options.delete(:dragdrop) ? 'filepicker-dragdrop' : 'filepicker'
-        @input_options = retrieve_legacy_filepicker_options(options)
-        @input_options['data-fp-apikey'] ||= ::Rails.application.config.filepicker_rails.api_key
-        @input_options.merge!(secure_filepicker) unless @input_options['data-fp-policy'].present?
-        @input_options['type'] = @type
-        @input_options
-      end
+      attr_reader :method, :object_name, :template
 
       def rails_greater_than_4_input
         tag = ActionView::Helpers::Tags::TextField.new(object_name, method, template, objectify_options(input_options))
@@ -81,25 +66,6 @@ module FilepickerRails
 
       def rails_greater_than_4?
         ::Rails.version.to_i >= 4
-      end
-
-      def filepicker_prefix
-        'data-fp-'
-      end
-
-      def retrieve_legacy_filepicker_options(options)
-        mappings = {}
-        FILEPICKER_OPTIONS_TO_DASHERIZE.each do |option|
-          mappings[option] = "#{filepicker_prefix}#{option.to_s.dasherize}"
-        end
-        FILEPICKER_OPTIONS_TO_CAMELIZE.each do |option|
-          mappings[option] = "#{filepicker_prefix}#{option.to_s.camelize(:lower)}"
-        end
-        Hash[options.map {|k, v| [mappings[k] || k, v] }]
-      end
-
-      def secure_filepicker
-        Policy.apply([:pick, :store], ['data-fp-policy', 'data-fp-signature'])
       end
   end
 end
