+++ lib/filepicker-rails.rb
@@ -1,6 +1,7 @@
 require "filepicker_rails/configuration"
 require "filepicker_rails/policy"
 require "filepicker_rails/engine"
+require "filepicker_rails/tag"
 
 module FilepickerRails
 end
